Fix some bad interactions between parcoord and crossfilter
1. Parcoord SHOULD NOT know about crossfilter
2. We should keep the strumming UX pattern as it is.

So, sliders PRODUCES a new dataset.
Strumming FILTERS on the current dataset.

But parcoord 'forget' to render brush when we call render
( the doc states it just renders the lines...)
so we need to recalculate the filtered datas when a slider
move.

In addition, brushing in parcoord takes the form of this interval :
[min, max[ .. this explains the 'weird behaviour' we have seen.

And then i have make super strange tweaks in the lib, due to
all the tweak we made :).. so we should fall back to
the original parcoord as of now i think that's the cleanest/most
consistent way...