Use position:absolute
That way, the element is taken out of the flow and put on the top
of the other element. This works 'by luck', because left
is unspecified. When it's the case, the user-agent is supposed
to look at right (which is 0), and then set the width according
to the content, which is what we want in that case.